Fix write bugs.
Firstly this fixes a failure to write because the directory entry was
not being marked as dirty.  Then it fixes writing the directory to the
wrong location, the sequence number to the wrong place within the
directory, corruption of the footer and writing the load, exec and
length as 24 bit values when they should have been 32 bit.