Fix filename ordering/conversion.
As ADFS directories rely on having the names in sorted order the
name_cmp function is design to work the same as the equivalent function
in Acorn's native ADFS.  It turns out, though, that this function is
only defined for the set of characters allowed in ADFS names, not the
set of host characters these are mapped to, thus keeping the directories
in memory in the host character set caused a problem.

This commit moves where the name conversion happens so the names are
converted into ADFS character set when they are received from the FUSE
library and converted back when given to the FUSE library.  The
adfs_dirent structure now keeps the names in the ADFS character set.