Fix rogue directories.
Some files with short names were randomly turning into directories.
This was because the byte of the filename that had bit 7 set to indicate
a directory was not being initialised to an ASCII character, with bit 7
clear, because copying the filename stopped at the zero terminator.
This change pads the copied filename to the end of allowable space with zeros.