Track pending deletes and pre-check free space map entries.
This change is aimed at avoiding a situation in which a delete is marked
as pending, waiting for the forget call from the kernel, and then when
the forget call arrives the free space map is full.  By then it is too
late to put the space being freed back to the original file for
consistency.

This change introduces a counter for pending deletes and before marking
a delete as pending checks that there is at least one free space map
entry for each already pending delete and the new one.